Implementing BLS-based mining.  BLS 12 381 operations: Removed exception block wrappers, improved const qualifications, added overloaded operators, RLP convertors.  Added RLP list extractor.  Reverted changes of replacing ECDSA by BLS in other parts of code.  Removed stake modifier from block header, moved to seal fields.  Fixed pre-compiled contract capitalizations.  Major adjustments to tests.

+++ libp2p/Session.cpp
@@ -355,7 +355,7 @@ void Session::disconnect(DisconnectReason _reason)
 	if (m_socket->ref().is_open())
 	{
 		RLPStream s;
-		prep(s, DisconnectPacket, 1) << (int)_reason;
+        prep(s, DisconnectPacket, 1) << (unsigned)(int)_reason;
 		sealAndSend(s, 0);
 	}
 	drop(_reason);
